src/wizard/requirements.txt
# SHARA Wizard of Oz Interface Dependencies

# GUI Framework
PyQt6>=6.4.0
PyQt6-WebEngine>=6.4.0

# Async support for Qt
qasync>=0.24.0

# WebSocket client
python-socketio[asyncio]>=5.8.0

# HTTP requests
requests>=2.31.0

# Image processing
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# JSON handling and data validation
pydantic>=2.0.0

# Logging and configuration
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-qt>=4.2.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Platform-specific dependencies
# Windows: pip install pywin32
# macOS: pip install pyobjc-framework-Cocoa
# Linux: sudo apt-get install python3-pyqt6.qtwebengine